refact(utils): Resolve circular dependency without duplicating any code
