+++ tns-core-modules/utils/mainthread-helper.ios.ts
@@ -1,0 +1,7 @@
+export function dispatchToMainThread(func: () => void) {
+    NSOperationQueue.mainQueue.addOperationWithBlock(func);
+}
+
+export function isMainThread(): Boolean {
+    return NSThread.isMainThread;
+}
