+++ tns-core-modules/utils/utils.ios.ts
@@ -159,14 +159,6 @@ export function openUrl(location: string): boolean {
     return false;
 }
 
-export function dispatchToMainThread(func: () => void) {
-    NSOperationQueue.mainQueue.addOperationWithBlock(func);
-}
-
-export function isMainThread(): Boolean {
-    return NSThread.isMainThread;
-}
-
 class UIDocumentInteractionControllerDelegateImpl extends NSObject implements UIDocumentInteractionControllerDelegate {
     public static ObjCProtocols = [UIDocumentInteractionControllerDelegate];
